# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoInfoRequest {
  /** YouTube video URL */
  url: string;
}

export interface VideoInfo {
  title: string;
  /** Duration in seconds */
  duration: number;
  /** Thumbnail URL */
  thumbnail: string;
  author: string;
}

/**
 * Audio or video format
 */
export type DownloadRequestFormat =
  (typeof DownloadRequestFormat)[keyof typeof DownloadRequestFormat];

export const DownloadRequestFormat = {
  mp3: "mp3",
  m4a: "m4a",
  wav: "wav",
  ogg: "ogg",
  flac: "flac",
  opus: "opus",
  mp4: "mp4",
  mkv: "mkv",
  webm: "webm",
  avi: "avi",
} as const;

export interface DownloadRequest {
  /** YouTube video URL */
  url: string;
  /** Audio or video format */
  format: DownloadRequestFormat;
}

export interface DownloadResponse {
  filename: string;
  downloadUrl: string;
  title: string;
  /** File size in bytes */
  fileSize?: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}