# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadRequestFormat } from "./downloadRequestFormat";

export interface DownloadRequest {
  /** YouTube video URL */
  url: string;
  /** Audio or video format */
  format: DownloadRequestFormat;
}