# lib/api-zod/src/generated/types/downloadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadResponse {
  filename: string;
  downloadUrl: string;
  title: string;
  /** File size in bytes */
  fileSize?: number;
}